Add write_parquet to dataframe
